Index.tsx refactor
